requirements/runtime.txt
requests==2.1.0

django==1.6.1
django-appconf==0.6         # provides graceful configuration of apps
django-compressor==1.3
django-filter==0.7          # provides generic filtering in API
djangorestframework==2.3.8
South==0.7.6

# Git provider requirements
GitPython==0.3.2.RC1            # access git repositories
